+++ semcore/tag/src/Tag.jsx
@@ -1,65 +1,61 @@
 import React from 'react';
 import createComponent, { Component, sstyled, Root } from '@semcore/core';
 import { Box } from '@semcore/flex-box';
-import resolveColor, { opacity } from '@semcore/utils/lib/color';
+import resolveColor, { opacity, light } from '@semcore/utils/lib/color';
 import addonTextChildren from '@semcore/utils/lib/addonTextChildren';
 import logger from '@semcore/utils/lib/logger';
 import CloseM from '@semcore/icon/Close/m';
 
 import style from './style/tag.shadow.css';
 
-function getPrimaryColors(color) {
-  const TEXT_BG_PRIMARY_MAP = {
-    'gray-500': 'gray-100',
-    'blue-500': 'blue-100',
-    'green-500': 'green-100',
-    'salad-500': 'salad-100',
-    'orange-500': 'orange-100',
-    'yellow-500': 'yellow-100',
-    'red-500': 'red-100',
-    'pink-500': 'pink-100',
-    'violet-500': 'violet-100',
-  };
+const textColorToBgColor = {
+  'gray-500': 'gray-100',
+  'blue-500': 'blue-100',
+  'green-500': 'green-100',
+  'salad-500': 'salad-100',
+  'orange-500': 'orange-100',
+  'yellow-500': 'yellow-100',
+  'red-500': 'red-100',
+  'pink-500': 'pink-100',
+  'violet-500': 'violet-100',
+};
+const textColorToHoveredTextColor = {
+  'gray-500': 'gray-50',
+  'blue-500': 'blue-50',
+  'green-500': 'green-50',
+  'salad-500': 'salad-50',
+  'orange-500': 'orange-50',
+  'yellow-500': 'yellow-50',
+  'red-500': 'red-50',
+  'pink-500': 'pink-50',
+  'violet-500': 'violet-50',
+};
+const textColorToBorderColor = {
+  'gray-500': 'gray-200',
+  'blue-500': 'blue-200',
+  'green-500': 'green-200',
+  'salad-500': 'salad-200',
+  'orange-500': 'orange-200',
+  'yellow-500': 'yellow-200',
+  'red-500': 'red-200',
+  'pink-500': 'pink-200',
+  'violet-500': 'violet-200',
+};
 
-  return {
-    colorBg: opacity(resolveColor(TEXT_BG_PRIMARY_MAP[color]), 0.5),
-    colorBgHover: resolveColor(TEXT_BG_PRIMARY_MAP[color]),
-    colorText: resolveColor(color),
-  };
-}
+const getPrimaryColors = (color) => ({
+  colorBg: opacity(resolveColor(textColorToBgColor[color]), 0.5) || light(color, 0.95),
+  colorBgHover: resolveColor(textColorToBgColor[color] || light(color, 0.8)),
+  colorText: resolveColor(color),
+});
 
-function getSecondaryColors(color) {
-  const TEXT_BG_HOVER_SECONDARY_MAP = {
-    'gray-500': 'gray-50',
-    'blue-500': 'blue-50',
-    'green-500': 'green-50',
-    'salad-500': 'salad-50',
-    'orange-500': 'orange-50',
-    'yellow-500': 'yellow-50',
-    'red-500': 'red-50',
-    'pink-500': 'pink-50',
-    'violet-500': 'violet-50',
-  };
-  const TEXT_BORDER_SECONDARY_MAP = {
-    'gray-500': 'gray-200',
-    'blue-500': 'blue-200',
-    'green-500': 'green-200',
-    'salad-500': 'salad-200',
-    'orange-500': 'orange-200',
-    'yellow-500': 'yellow-200',
-    'red-500': 'red-200',
-    'pink-500': 'pink-200',
-    'violet-500': 'violet-200',
-  };
-  return {
-    colorBg: resolveColor('white'),
-    colorBgHover: resolveColor(TEXT_BG_HOVER_SECONDARY_MAP[color]),
-    colorText: resolveColor(color),
-    colorBorder: resolveColor(TEXT_BORDER_SECONDARY_MAP[color]),
-  };
-}
+const getSecondaryColors = (color) => ({
+  colorBg: resolveColor('white'),
+  colorBgHover: resolveColor(textColorToHoveredTextColor[color] || light(color, 0.95)),
+  colorText: resolveColor(color),
+  colorBorder: resolveColor(textColorToBorderColor[color] || light(color, 0.5)),
+});
 
-const MAP_OLD_THEME = {
+const legacyThemeRecommendedMigration = {
   primary: {
     muted: 'gray-500',
     info: 'blue-500',
@@ -70,14 +66,13 @@ const MAP_OLD_THEME = {
   secondary: {
     muted: 'gray-50',
   },
-  undefined: {},
 };
 
 class RootTag extends Component {
   static displayName = 'Tag';
   static style = style;
   static defaultProps = {
-    theme: 'secondary',
+    theme: 'primary',
     color: 'gray-500',
     size: 'm',
   };
@@ -88,7 +83,7 @@ class RootTag extends Component {
     logger.warn(
       props.use,
       `Property 'use' is deprecated, replace property to "theme='${props.use}' color='${
-        MAP_OLD_THEME[props.use][props.theme]
+        legacyThemeRecommendedMigration[props.use]?.[props.theme]
       }'"`,
       props['data-ui-name'] || Tag.displayName,
     );
@@ -104,13 +99,9 @@ class RootTag extends Component {
     const { Children, styles, theme, color, interactive, disabled, addonLeft, addonRight } =
       this.asProps;
 
-    let colors = {};
-    if (theme === 'primary') {
-      colors = getPrimaryColors(color);
-    }
-    if (theme === 'secondary') {
-      colors = getSecondaryColors(color);
-    }
+    const colors = theme !== 'primary' ? getSecondaryColors(color) : getPrimaryColors(color);
+
+    console.log(colors);
 
     return sstyled(styles)(
       <STag