+++ semcore/utils/src/color.ts
@@ -146,3 +146,113 @@ export default function resolveColor(color?: string): string {
   }
   return color;
 }
+
+const lightMemo: { [input: string]: string } = {};
+export const light = (color?: string, factor: number = 1) => {
+  if (!color) return '';
+  const inputKey = `${color}_${factor}`;
+  if (lightMemo[inputKey]) return lightMemo[inputKey];
+
+  let r = 0,
+    g = 0,
+    b = 0,
+    a = 1;
+  if (color.startsWith('#')) {
+    if (color.length !== 4 && color.length !== 7) {
+      throw new Error(`Got invalid color ${color}`);
+    }
+    const partSize = color.length === 4 ? 1 : 2;
+    r = parseInt(color.substring(1, 1 + partSize), 16);
+    g = parseInt(color.substring(1 + partSize, 1 + partSize * 2), 16);
+    b = parseInt(color.substring(1 + partSize * 2, 1 + partSize * 3), 16);
+    a = 1; // not handled for now
+  } else if (color.startsWith('rgb(')) {
+    const trimmed = color.substring('rgb('.length, color.length - ')'.length);
+    const parts = trimmed.split(',').map((part) => parseInt(part.trim(), 10));
+    r = parts[0];
+    g = parts[1];
+    b = parts[2];
+    a = 1;
+  } else if (color.startsWith('rgba(')) {
+    const trimmed = color.substring('rgba('.length, color.length - ')'.length);
+    const parts = trimmed.split(',').map((part) => parseInt(part.trim(), 10));
+    r = parts[0];
+    g = parts[1];
+    b = parts[2];
+    a = parts[2];
+  } else {
+    throw new Error(`Got invalid color ${color}`);
+  }
+
+  let hsl = rgbToHsl(r, g, b);
+
+  hsl[2] += (1 - hsl[2]) * factor;
+
+  const rgb = hslToRgb(...hsl);
+
+  requestAnimationFrame(() => {
+    console.log(lightMemo, [r, g, b, a], hsl);
+  });
+  return (lightMemo[inputKey] = `rgba(${rgb[0]}, ${rgb[1]}, ${rgb[2]}, ${a})`);
+};
+
+const rgbToHsl = (r: number, g: number, b: number) => {
+  (r /= 255), (g /= 255), (b /= 255);
+
+  var max = Math.max(r, g, b),
+    min = Math.min(r, g, b);
+  var h,
+    s,
+    l = (max + min) / 2;
+
+  if (max == min) {
+    h = s = 0;
+  } else {
+    var d = max - min;
+    s = l > 0.5 ? d / (2 - max - min) : d / (max + min);
+
+    switch (max) {
+      case r:
+        h = (g - b) / d + (g < b ? 6 : 0);
+        break;
+      case g:
+        h = (b - r) / d + 2;
+        break;
+      case b:
+        h = (r - g) / d + 4;
+        break;
+    }
+
+    h /= 6;
+  }
+
+  return [h, s, l] as [h: number, s: number, l: number];
+};
+
+const hslToRgb = (h: number, s: number, l: number) => {
+  var r = 0,
+    g = 0,
+    b = 0;
+
+  if (s == 0) {
+    r = g = b = l;
+  } else {
+    function hue2rgb(p: number, q: number, t: number) {
+      if (t < 0) t += 1;
+      if (t > 1) t -= 1;
+      if (t < 1 / 6) return p + (q - p) * 6 * t;
+      if (t < 1 / 2) return q;
+      if (t < 2 / 3) return p + (q - p) * (2 / 3 - t) * 6;
+      return p;
+    }
+
+    var q = l < 0.5 ? l * (1 + s) : l + s - l * s;
+    var p = 2 * l - q;
+
+    r = hue2rgb(p, q, h + 1 / 3);
+    g = hue2rgb(p, q, h);
+    b = hue2rgb(p, q, h - 1 / 3);
+  }
+
+  return [r * 255, g * 255, b * 255] as [r: number, g: number, b: number];
+};